+++ src/app/PageWarmup.h
@@ -17,6 +17,7 @@ class PageWarmup : public QObject {
   Q_PROPERTY(int currentPage READ currentPage WRITE setCurrentPage NOTIFY
                  currentPageChanged)
   Q_PROPERTY(QSize tileSize READ tileSize WRITE setTileSize NOTIFY tileSizeChanged)
+  Q_PROPERTY(bool paused READ paused WRITE setPaused NOTIFY pausedChanged)
 
 public:
   explicit PageWarmup(QObject *parent = nullptr);
@@ -31,10 +32,14 @@ public:
   [[nodiscard]] QSize tileSize() const { return m_tileSize; }
   void setTileSize(QSize size);
 
+  [[nodiscard]] bool paused() const { return m_paused; }
+  void setPaused(bool paused);
+
 signals:
   void documentChanged();
   void currentPageChanged();
   void tileSizeChanged();
+  void pausedChanged();
 
 private slots:
   void onDocumentStatus();
@@ -50,4 +55,5 @@ private:
   QString m_ownPath;
   int m_currentPage{-1};
   QSize m_tileSize{720, 960};
+  bool m_paused{false};
 };